frontend: move api wrapeprs to api folder
Small organization thing, that's all
