+++ frontend/src/api/hydra.js
@@ -1,4 +1,4 @@
-import config from './config';
+import config from '../config';
 
 class HydraAPI {
   static getConsent(id) {
